test(#656): split photo-search route/contract tests by topic (≤200 lines)
CodeRabbit findings 2+3 (#739, confirmed): test_photo_search_route.py
(421 lines) and test_photo_search_contract.py (215 lines) both exceeded
the repo's 200-line test file budget.

Unit (agent/tests/unit/):
- photo_search_route_fixtures.py — shared app_/body_/settings_ builders,
  FunctionModel fakes, and the UsageRepo fake (not collected as tests).
- test_photo_search_route.py (70 lines) — core contract: happy path,
  platform-outage degrade, confirm endpoint.
- test_photo_search_route_validation.py (42 lines) — image format/decode/
  size checks.
- test_photo_search_route_byok.py (109 lines) — BYOK login gate + BYOK
  fallback/success usage attribution.
- test_photo_search_route_quota.py (168 lines) — budget breaker, quota
  key/tiers/guidance, usage-scope attribution, and the two new regression
  tests for the quota-ordering fix (previous commit).
- test_photo_search_anonymous_tier.py now imports app_/body_/settings_
  from the shared fixtures module instead of a specific test file.

Integration (agent/tests/integration/):
- photo_search_contract_fixtures.py — shared model/app/body builders.
- test_photo_search_contract.py (58 lines) — resolve/degrade/layer-2 core
  contract against the named image fixtures.
- test_photo_search_contract_byok.py (57 lines) — BYOK-first resolution
  and BYOK-failure fallback, end to end.
- test_photo_search_contract_quota.py (29 lines) — quota tiers and
  guidance premises.

Each split file now has one topic it can name; no file crosses 200 lines.
Total test count is unchanged except for the two new quota-ordering
regression tests from the previous commit.
